Chatbot_GptVersion/requirements.txt
streamlit
langchain
faiss-cpu
PyPDF2
google-generativeai
requests
sentence-transformers
